refactor(user): update ID and token invalidation method return types
- Changed `id` field type from `Long` to `Integer` in `UserResponse` record and `UserManagementUseCase` interface to align with domain model.
- Updated return type of `invalidateUsedToken` methods from `Mono<Void>` to `Mono<String>` to provide a clearer response message upon completion.

+++ services/bank-user/src/main/java/am/banking/system/user/application/port/in/UserManagementUseCase.java
@@ -10,7 +10,7 @@ import reactor.core.publisher.Mono;
  * Time: 18:43:28
  */
 public interface UserManagementUseCase {
-    Mono<Void> updateUserAccountState(Long id);
+    Mono<Void> updateUserAccountState(Integer id);
     Mono<Result<UserResponse>> getUserByUsername(String username);
     Mono<Result<UserResponse>> getUserByEmail(String email);
 }
